requirements.txt
# Production dependencies (umbrella)

# Core
aiohttp==3.10.5
selenium==4.15.2
requests==2.31.0
python-telegram-bot==20.7

# Parsing / data
beautifulsoup4==4.12.2
lxml==5.2.2
pandas>=2.2.0
html5lib==1.1  # opcional

# Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
structlog==23.2.0

# Network capture (for Betburger/Surebet tokens)
selenium-wire==5.1.0
setuptools>=68.0.0
packaging>=23.2

# OCR (opcional, requiere tesseract en el sistema)
pytesseract==0.3.10
Pillow==10.3.0

# Nota: paquetes de desarrollo/pruebas están en requirements-dev.txt
